Paper Examples/deep_operators/requirements.txt
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
comm==0.1.3
configmypy==0.1.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
DeepXDE==1.9.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.18
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.1
neuraloperator==0.2.0
nose==1.3.7
numcodecs==0.11.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==23.5.9
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sacred==0.8.4
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.25.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
tensorly==0.8.1
tensorly-torch==0.4.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
wandb==0.15.3
wcwidth==0.2.6
wrapt==1.15.0
zarr==2.14.2
zipp==3.15.0
